requirements.txt
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
countryinfo==0.1.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.5
geopandas==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.8
logistro==1.1.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
narwhals==1.46.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
plotly-express==0.4.1
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
python3-discogs-client==2.8
pytz==2025.2
pywin32==310
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.16.0
seaborn==0.13.2
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.5
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
